Remove incorrect information about `git stash'
I said that if you don't use the `--index' option to `git stash pop'
or `git stash apply' then git will lose the changes that were in your
index when you created the stash.  This is incorrect.

Git does keep these changes, it just puts them in the working
directory and not the index.  Since it doesn't lose your changes I
don't think it's necessary to teach that you should use `--index'.